Move export_helper into shared, add PresenterClass type.
