+++ openslides_backend/action/actions/meeting/clone.py
@@ -13,10 +13,10 @@ from openslides_backend.shared.interfaces.event import Event, EventType
 from openslides_backend.shared.patterns import fqid_from_collection_and_id
 from openslides_backend.shared.schema import id_list_schema, required_id_schema
 
+from ....shared.export_helper import export_meeting
 from ...util.default_schema import DefaultSchema
 from ...util.register import register_action
 from ...util.typing import ActionData
-from .export_helper import export_meeting
 from .import_ import ONE_ORGANIZATION_ID, MeetingImport
 
 updatable_fields = [
