+++ openslides_backend/presenter/check_database.py
@@ -3,11 +3,11 @@ from typing import Any, Dict, List
 import fastjsonschema
 from datastore.shared.util import DeletedModelsBehaviour
 
-from ..action.actions.meeting.export_helper import export_meeting
 from ..models.checker import Checker, CheckException
 from ..permissions.management_levels import OrganizationManagementLevel
 from ..permissions.permission_helper import has_organization_management_level
 from ..shared.exceptions import PermissionDenied
+from ..shared.export_helper import export_meeting
 from ..shared.schema import optional_id_schema, schema_version
 from .base import BasePresenter
 from .presenter import register_presenter
