+++ openslides_backend/presenter/export_meeting.py
@@ -2,10 +2,10 @@ from typing import Any, Dict
 
 import fastjsonschema
 
-from ..action.actions.meeting.export_helper import export_meeting
 from ..permissions.management_levels import OrganizationManagementLevel
 from ..permissions.permission_helper import has_organization_management_level
 from ..shared.exceptions import PermissionDenied
+from ..shared.export_helper import export_meeting
 from ..shared.schema import required_id_schema, schema_version
 from .base import BasePresenter
 from .presenter import register_presenter
